Fix bug with missing close div tag in html preview
Symptoms

When viewing the summary Preview.html generated by deliver, if you had
many languages, each language would appear to have more and more
padding applied to the width. This meant that languages that were
displayed later in the preview would appear in a much narrower column
rather than the full width of the screen. It made them hard to read.

Fix

There was a missing </div> tag at the end of the last row of
screenshots per language. This meant that the padding from earlier divs
was being reapplied for each subsequent language which led to gradually
narrower spacing.

As a precaution, a check was added around the last closing div to make
sure that we actually added some screenshots.